登录错误次数bug 登录记录日志异步

+++ src/main/java/com/seezoon/admin/modules/sys/service/LoginSecurityService.java
@@ -8,6 +8,7 @@ import org.springframework.data.redis.core.ValueOperations;
 import org.springframework.stereotype.Service;
 import org.springframework.util.Assert;
 
+import com.alibaba.fastjson.support.spring.GenericFastJsonRedisSerializer;
 import com.seezoon.boot.common.service.BaseService;
 
 
@@ -15,7 +16,7 @@ import com.seezoon.boot.common.service.BaseService;
 public class LoginSecurityService extends BaseService {
 	
 	@Resource(name="redisTemplate")
-	private ValueOperations<String, Long> valueOperations;
+	private ValueOperations<String, Integer> valueOperations;
 	private String LOCK_PREFIX = "login_cnt_";
 
 	public void unLock(String loginName) {
@@ -23,9 +24,9 @@ public class LoginSecurityService extends BaseService {
 		valueOperations.getOperations().delete(LOCK_PREFIX + loginName);
 	}
 
-	public Long getLoginFailCount(String loginName) {
+	public Integer getLoginFailCount(String loginName) {
 		try {
-			Long cnt = valueOperations.get(LOCK_PREFIX + loginName);
+			Integer cnt = valueOperations.get(LOCK_PREFIX + loginName);
 			return null == cnt ? 0:cnt;
 		} finally {
 			valueOperations.getOperations().expire(LOCK_PREFIX + loginName, 24, TimeUnit.HOURS);
@@ -37,6 +38,9 @@ public class LoginSecurityService extends BaseService {
 	}
 	public Long incrementLoginFailTimes(String loginName) {
 		try {
+			//实际存放在redis 中是integer的，超过integer才会到long spring data increment 
+			//在GenericFastJsonRedisSerializer/GenericJackson2JsonRedisSerializer 下的bug
+			//测试RedisAtomicLong 也会有类型问题，为了改动小故采用integer 正确使用是通过increment 获取值，而不是直接get
 			Long increment = valueOperations.increment(LOCK_PREFIX + loginName, 1);
 			return increment;
 		} finally {
